Extract validation into separate module w/ tests
